api, jwt, swagger

+++ MicroServiceProject/Source/Core/MP.Core.Application/Repositories/IGenericRepository.cs
@@ -12,7 +12,7 @@ namespace MP.Core.Application.Repositories
     {
         public Task<List<T>> GetAll();
         public Task<int> GetAllCount();
-        public Task<T> GetById(int id);        
+        public Task<T> GetById(long id);        
         public Task<T> Insert(T model);
         public Task<List<T>> MultipleInsert(List<T> model);
         public Task<T> Update(T model);
